fix session overriding fetchSize of driver config

+++ Neo4j.Driver/Neo4j.Driver.Tests.TestBackend/Protocol/Driver/NewDriver.cs
@@ -29,7 +29,7 @@ namespace Neo4j.Driver.Tests.TestBackend
             [JsonIgnore]
             public bool ModifiedTrustedCertificates = false;
 
-            public int? fetchSize;
+            public long? fetchSize;
 
             public string[] trustedCertificates
             {
